requirements/prod.txt
Django==5.1.*
djangorestframework==3.15.*
tika==2.6.*
psycopg2-binary
pdfminer.six